refactor: move sightings and vehicles queries to queries/
Moves the Sequelize data access in services/sightings.ts and
services/vehicles.ts to src/queries/, so all database queries live in
one place (matching the queries/ convention used by the newer
features). Pure file move plus import updates; no behaviour change.
