+++ src/queries/sightings.ts
