+++ src/queries/vehicles.ts
